Fix compilation.

+++ crawl-ref/source/religion.cc
@@ -1090,7 +1090,7 @@ static void _inc_penance(int val)
 static void _inc_gift_timeout(int val)
 {
     you.gift_timeout += val;
-    you.gift_timeout = std::min(200, you.gift_timeout);
+    you.gift_timeout = std::min((uint8_t)200, you.gift_timeout);
 }
 
 // These are sorted in order of power.
